Rename file to match rest of v2 quirks
